VCV-185: Create tab for sites list in review section
