+++ src/features/review/components/sites-list/review-visit-coverage-badge.tsx
@@ -1,30 +1,32 @@
 import { Spinner } from '@/components/ui/spinner';
 
-interface CompletenessBoxProps {
-    percentage: number;
+function getVisitCoverageColor(
+    visitedPercentage: number,
+    highThreshold: number,
+    mediumThreshold: number,
+): string {
+    if (visitedPercentage >= highThreshold)
+        return 'bg-success/10 text-success border-success/50';
+
+    if (visitedPercentage >= mediumThreshold)
+        return 'bg-warning/20 text-warning-foreground border-warning/50';
+
+    return 'bg-destructive/20 text-destructive border-destructive/50';
+}
+
+interface ReviewVisitCoverageBadgeProps {
+    visitedPercentage: number;
     isLoading?: boolean;
     highThreshold?: number;
     mediumThreshold?: number;
 }
 
-function getCompletenessColor(
-    percentage: number,
-    highThreshold: number,
-    mediumThreshold: number,
-): string {
-    if (percentage >= highThreshold)
-        return 'bg-success/10 text-success border-success/50';
-    if (percentage >= mediumThreshold)
-        return 'bg-warning/20 text-warning-foreground border-warning/50';
-    return 'bg-destructive/20 text-destructive border-destructive/50';
-}
-
-export function CompletenessBox({
-    percentage,
+export default function ReviewVisitCoverageBadge({
+    visitedPercentage,
     isLoading = false,
     highThreshold = 80,
     mediumThreshold = 50,
-}: CompletenessBoxProps) {
+}: ReviewVisitCoverageBadgeProps) {
     if (isLoading) {
         return (
             <span className="bg-muted text-muted-foreground border-border rounded border px-2 py-1">
@@ -32,11 +34,16 @@ export function CompletenessBox({
             </span>
         );
     }
+
     return (
         <span
-            className={`rounded border px-2 py-1 ${getCompletenessColor(percentage, highThreshold, mediumThreshold)}`}
+            className={`rounded border px-2 py-1 ${getVisitCoverageColor(
+                visitedPercentage,
+                highThreshold,
+                mediumThreshold,
+            )}`}
         >
-            {Math.round(percentage)}%
+            {Math.round(visitedPercentage)}%
         </span>
     );
 }
